Move clip_off_prefix to the common ena_std_lib module.
