+++ ena_objects/ena_run.py
@@ -3,8 +3,7 @@ from typing import List, Dict
 from pandas import DataFrame
 
 from ena_objects.characteristic import IsaBase
-from ena_objects.ena_experiment import clip_off_prefix
-from ena_objects.ena_std_lib import get_assay_sample_associations
+from ena_objects.ena_std_lib import get_assay_sample_associations, clip_off_prefix
 
 
 class DataFileComment(IsaBase):
